npmjs_exercise/analysis_server/requirements.txt
djangorestframework
expiringdict
requests